feat: implement MySQL storage for student creation and update student…

+++ internal/types/types.go
@@ -1,7 +1,7 @@
 package types
 
 type Student struct {
-	ID    int    `json:"id"  validate:"-"`
+	ID    int64  `json:"id"  validate:"-"`
 	Name  string `json:"name" validate:"required"`
 	Email string `json:"email" validate:"required,email"`
 	Class string `json:"class" validate:"required"`
